reverts payments method keyboard PR
