+++ src/CONST.js
@@ -790,7 +790,6 @@ const CONST = {
         INVITE: 'invite',
         LEAVE_ROOM: 'leaveRoom',
     },
-    IOS_KEYBOARD_SPACE_OFFSET: -30,
 };
 
 export default CONST;
