+++ src/components/PasswordPopover/PasswordPopover.js
@@ -1,30 +1,41 @@
 import {TouchableOpacity, View} from 'react-native';
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
-import styles from '../../styles/styles';
-import Text from '../Text';
-import Popover from '../Popover';
-import withLocalize, {withLocalizePropTypes} from '../withLocalize';
-import compose from '../../libs/compose';
-import withWindowDimensions from '../withWindowDimensions';
-import TextInput from '../TextInput';
-import KeyboardSpacer from '../KeyboardSpacer';
-import {propTypes as passwordPopoverPropTypes, defaultProps as passwordPopoverDefaultProps} from './passwordPopoverPropTypes';
+import styles from '../styles/styles';
+import Text from './Text';
+import Popover from './Popover';
+import withLocalize, {withLocalizePropTypes} from './withLocalize';
+import compose from '../libs/compose';
+import withWindowDimensions from './withWindowDimensions';
+import TextInput from './TextInput';
 
 const propTypes = {
-    /** Whether we should wait before focusing the TextInput, useful when using transitions on Android */
-    shouldDelayFocus: PropTypes.bool,
+    /** Is the popover currently showing? */
+    isVisible: PropTypes.bool.isRequired,
 
-    ...passwordPopoverPropTypes,
+    /** Function that gets called when the user closes the modal */
+    onClose: PropTypes.func.isRequired,
+
+    /** Where the popover should be placed */
+    anchorPosition: PropTypes.shape({
+        top: PropTypes.number,
+        left: PropTypes.number,
+    }).isRequired,
+
+    /** Function that gets called when the user clicks the delete / make default button */
+    onSubmit: PropTypes.func,
+
+    /** The text that should be displayed in the submit button */
+    submitButtonText: PropTypes.string,
     ...withLocalizePropTypes,
 };
 
 const defaultProps = {
-    shouldDelayFocus: false,
-    ...passwordPopoverDefaultProps,
+    onSubmit: () => {},
+    submitButtonText: '',
 };
 
-class BasePasswordPopover extends Component {
+class PasswordPopover extends Component {
     constructor(props) {
         super(props);
 
@@ -80,7 +91,6 @@ class BasePasswordPopover extends Component {
                         onSubmitEditing={() => this.props.onSubmit(this.state.password)}
                         style={styles.mt3}
                         autoFocus
-                        shouldDelayFocus={this.props.shouldDelayFocus}
                     />
                     <TouchableOpacity
                         onPress={() => this.props.onSubmit(this.state.password)}
@@ -95,15 +105,14 @@ class BasePasswordPopover extends Component {
                         </Text>
                     </TouchableOpacity>
                 </View>
-                <KeyboardSpacer />
             </Popover>
         );
     }
 }
 
-BasePasswordPopover.propTypes = propTypes;
-BasePasswordPopover.defaultProps = defaultProps;
+PasswordPopover.propTypes = propTypes;
+PasswordPopover.defaultProps = defaultProps;
 export default compose(
     withWindowDimensions,
     withLocalize,
-)(BasePasswordPopover);
+)(PasswordPopover);
