--- src/components/PasswordPopover/index.android.js
@@ -1,18 +1,0 @@
-import React, {forwardRef} from 'react';
-import BasePasswordPopover from './BasePasswordPopover';
-import {propTypes, defaultProps} from './passwordPopoverPropTypes';
-
-const PasswordPopover = forwardRef((props, ref) => (
-    <BasePasswordPopover
-        // eslint-disable-next-line react/jsx-props-no-spreading
-        {...props}
-        ref={ref}
-        shouldDelayFocus
-    />
-));
-
-PasswordPopover.propTypes = propTypes;
-PasswordPopover.defaultProps = defaultProps;
-PasswordPopover.displayName = 'PasswordPopover';
-
-export default PasswordPopover;
