+++ src/components/TextInput/baseTextInputPropTypes.js
@@ -64,9 +64,6 @@ const propTypes = {
 
     /** Callback to update the value on Form when input is used in the Form component. */
     onInputChange: PropTypes.func,
-
-    /** Whether we should wait before focusing the TextInput, useful when using transitions  */
-    shouldDelayFocus: PropTypes.bool,
 };
 
 const defaultProps = {
@@ -96,7 +93,6 @@ const defaultProps = {
     hint: '',
     prefixCharacter: '',
     onInputChange: () => {},
-    shouldDelayFocus: false,
 };
 
 export {propTypes, defaultProps};
