+++ src/pages/home/sidebar/SidebarLinks.js
@@ -15,6 +15,7 @@ import ROUTES from '../../../ROUTES';
 import Icon from '../../../components/Icon';
 import Header from '../../../components/Header';
 import OptionsList from '../../../components/OptionsList';
+import KeyboardSpacer from '../../../components/KeyboardSpacer';
 import * as Expensicons from '../../../components/Icon/Expensicons';
 import AvatarWithIndicator from '../../../components/AvatarWithIndicator';
 import * as OptionsListUtils from '../../../libs/OptionsListUtils';
@@ -328,6 +329,7 @@ class SidebarLinks extends React.Component {
                     optionMode={this.props.priorityMode === CONST.PRIORITY_MODE.GSD ? 'compact' : 'default'}
                     onLayout={App.setSidebarLoaded}
                 />
+                <KeyboardSpacer />
             </View>
         );
     }
