+++ src/styles/StyleUtils.js
@@ -483,28 +483,6 @@ function convertToLTR(text) {
     return `\u2066${text}`;
 }
 
-/**
- * Checks to see if the iOS device has safe areas or not
- *
- * @param {Number} windowWidth
- * @param {Number} windowHeight
- * @returns {Boolean}
- */
-function hasSafeAreas(windowWidth, windowHeight) {
-    const heightsIphonesWithNotches = [812, 896, 844, 926];
-    return _.contains(heightsIphonesWithNotches, windowHeight) || _.contains(heightsIphonesWithNotches, windowWidth);
-}
-
-/**
- * Get variable keyboard height as style
- * @param {Number} keyboardHeight
- * @returns {Object}
- */
-function getHeight(keyboardHeight) {
-    return {
-        height: keyboardHeight,
-    };
-}
 export {
     getAvatarSize,
     getAvatarStyle,
@@ -535,6 +513,4 @@ export {
     getPaddingLeft,
     getReportListAnimationStyle,
     convertToLTR,
-    hasSafeAreas,
-    getHeight,
 };
